Bug-fixes for inter-thread forward cellular analysis.
Changes in this and the last bunch of commits ensure:
1. FunctionDefinitionInfo.getCallersOfThis now correctly uses
getReachableCallStatementsInclusive instead of
getLexicallyEnclosedCallStatements. This ensures that call-sites from
functions that are not reachable from the main function are being
ignored. This, in turn, has now fixed an issue with visit of
ParameterDeclaration in DataFlowAnalysis.
2. Fixed the order in which values are printed for flow-facts, by using
the correct Comparator method in CellularFlowMap.getString.
3. The IN flow-fact object for those nodes that are part of singleton
SCCs, was not being regenerated during incremental analysis. Fixed this
bug.
4. A separate method "processNonSCCNodeOptimized" has been created for
handling the nodes that are part of singleton SCCs.
5. Fixed the issue where "inChanged" was not being dependent upon
equality of oldIN and newIN for the case when a new IN object was being
used for a node during its first processing during incremental update.
6. Fixed an issue with "stage" of the worklist during "recreate" -- now
it is being correctly reset to NONBARRIER.
7. To prevent processing of any SCC node more than once, we now maintain
only one stage in the worklist -- all barriers are now treated as
non-barriers when adding an element to the worklists.
8. Now we do not maintain PTA information for the FieldCell for those
arrays whose elements are not array/pointer types.
9. Fixed the usage of the worklist in the second pass of an SCC.
10. Fixed the same for barriers.
11. Now we do not process any node more than once in the second part of
the first pass over an SCC.
12. Fixed the issue where changes to the global scope was not adding
entry points of main to the set of seed nodes.
13. Fixed the order of memoization of call-statements and DFD-symbols.

Next, we need to fix the issue with those PTA sets that do not correctly
model the NULL element alongside non-NULL elements.